+++ changelog.md
@@ -2,12 +2,12 @@
 * Fixed issue in [Subscription DSL](https://occurrent.org/documentation#subscription-dsl) when using "subscribe" functions with a single event type different from the "base event type", i.e. this didn't work in previous version in Java:
   ```java                
   // GameEvent is the "base event type"
-  Subscriptions<GameEvent> subsctriptions = new Subscriptions<>(..);
+  Subscriptions<GameEvent> subscriptions = new Subscriptions<>(..);
   
   // GameStarted has GameEvent as parent, the following didn't compile in version 0.17.0 
-  subscriptions.subscribe("ikk", GameStarted.class, gameStarted -> System.out.println("gameStarted: " + gameStarted));
+  subscriptions.subscribe("mysubscription", GameStarted.class, gameStarted -> System.out.println("gameStarted: " + gameStarted));
   ```
-* Using slf4j-api and not logback-classic in seUsing slf4j-api and not logback-classic in several modules that accidentally brought logback in a compile dependencyveral modules that accidentally brought logback in a compile dependency.                                                                                 
+* Using slf4j-api and not logback-classic in several modules that accidentally brought logback in as a compile time dependency.                                                                                 
 * Upgraded slf4j-api from 2.0.5 to 2.0.12
 
 ### 0.17.0 (2024-01-19)