+++ changelog.md
@@ -1,4 +1,4 @@
-### Next version
+### 0.17.1 (2024-02-23)
 * Fixed issue in [Subscription DSL](https://occurrent.org/documentation#subscription-dsl) when using "subscribe" functions with a single event type different from the "base event type", i.e. this didn't work in previous version in Java:
   ```java                
   // GameEvent is the "base event type"